refactor(admin): AdminSeances sous 300 lignes — éclatement complet (#G1 ≤300)
AdminSeances.vue : 666 -> 221. Composable useAdminSeances (état + filtres + load/cache/
refresh + handlers) ; sous-composants SeancesFilters (v-model des 4 filtres) et
SeancesList (grille de cartes + helpers de présentation). La vue orchestre header +
états + pagination + sous-composants. CSS distribué verbatim (header bespoke conservé).

Dette préexistante conservée (parité) : filtre statut no-op (bindé mais non envoyé à
l'API), viewSeanceDetails/enableVisio = stubs TODO — comportement inchangé, documenté.

Tests : useAdminSeances + SeancesFilters + SeancesList. 15/15 verts. vite build vert.
Fichiers ≤300.
